+++ tests/testthat/test-discontinuous.R
@@ -14,6 +14,15 @@ test_that("Staircase: tolerance", {
 test_that("Staircase: f(a) and f(b) of opposite signs", {
   testthat::expect_false(sign(res$f.a) == sign(res$f.b))
 })
+# Repeat for -staircase
+neg_staircase <- function(x) -staircase(x)
+res <- itp(neg_staircase, c(-1, 1), epsilon = epsilon)
+test_that("-Staircase: tolerance", {
+  testthat::expect_lte(res$b - res$a , 2 * epsilon)
+})
+test_that("-Staircase: f(a) and f(b) of opposite signs", {
+  testthat::expect_false(sign(res$f.a) == sign(res$f.b))
+})
 
 # Trunc poly
 
@@ -23,5 +32,14 @@ test_that("Truncpoly: tolerance", {
   testthat::expect_lte(res$b - res$a , 2 * epsilon)
 })
 test_that("Truncpoly: f(a) and f(b) of opposite signs", {
+  testthat::expect_false(sign(res$f.a) == sign(res$f.b))
+})
+# Repeat for -truncpoly
+neg_truncpoly <- function(x) -truncpoly(x)
+res <- itp(neg_truncpoly, c(-1, 1), epsilon = epsilon)
+test_that("-Truncpoly: tolerance", {
+  testthat::expect_lte(res$b - res$a , 2 * epsilon)
+})
+test_that("-Truncpoly: f(a) and f(b) of opposite signs", {
   testthat::expect_false(sign(res$f.a) == sign(res$f.b))
 })