+++ NEWS.md
@@ -1,5 +1,5 @@
 # itp 1.0.0.9000
 
-## New features
-
 ## Bug fixes and minor improvements
+
+* A bug in the part of `itp()` in which the bracketing interval has been fixed. It caused incorrect results in cases where the function is locally decreasing.